+++ src/lib.rs
@@ -18,8 +18,8 @@ pub mod utils;
 #[cfg(not(feature = "utils"))]
 mod utils;
 
-pub use errors::{ItemCountMismatch, KeyCollision, OneOf2, ValueCollision};
-pub use traits::{TryCollectEx, TryExtend, TryExtendSafe, TryFromIterator, TryUnzip};
+pub use errors::*;
+pub use traits::*;
 
 #[cfg(feature = "arrayvec")]
 pub use errors::ExceedsCapacity;